pyproject.toml
[tool.poetry]
name = "strawberry-django-auth"
version = "0.372.0"
description = "Graphql authentication system with Strawberry for Django."
license = "MIT"
authors = ["Nir.J Benlulu <nrbnlulu@gmail.com>"]
maintainers = ["Nir.J Benlulu <nrbnlulu@gmail.com>"]
readme = "README.md"
classifiers = [
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Framework :: Django',
    'Framework :: Django :: 3.2',
    'Framework :: Django :: 4.1',
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
packages = [
    { include = "gqlauth" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
Django = ">=3.2,<4.2"
PyJWT = ">=2.6.0,<3.0"
strawberry-graphql-django = { version = ">=0.2.5,<4.0", allow-prereleases = true }
Pillow = "^9.2.0"
strawberry-graphql = ">=0.128,<0.171.0"
strawberry-django-plus = "^1.25.2"

[tool.poetry.dev-dependencies]
cryptography = "^39.0"
coverage = "^7.0"
pytest = "^7.2"
pytest-cov = "^4.0"
pytest-django = "^4.0.0"
types-cryptography = "^3.3.23"
django-mock-queries = "^2.1.7"
types-mock = "^4.0.15"
types-jwt = "^0.1.0"
types-pkg-resources = "^0.1.0"
mypy = "^0.961"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}
django-stubs-ext = "^0.7"
mkdocs = "^1.3.0"
mkdocs-markdownextradata-plugin = "^0.2.5"
mkdocs-material = "^8.5.3"
mkdocs-minify-plugin = "^0.5.0"
pymdown-extensions = "^9.5"
Markdown = "^3.4.1"
Faker = "^15.3.4"
pre-commit = "^2.21.0"
pytest-asyncio = "^0.20.3"
pydoc-markdown = {extras = ["novella"], version = "^4.6.3"}
gql = {extras = ["all"], version = "^3.4.0"}
setuptools = "^65.6.3"
channels = {extras = ["daphne"], version = "^4.0.0"}


[tool.pytest.ini_options]
django_find_project = false
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "testproject.settings"
python_files = "tests.py test_*.py *_tests.py"
asyncio_mode = "auto"
markers = [
    "default_user: uses the default settings, a user with email address and default USERNAME_FIELD",
    "settings_b: uses a custom user model, not captcha validation as well as some other non-default settings"
]

[tool.poetry.urls]
"Homepage" = "https://github.com/nrbnlulu/strawberry-django-auth"
"Documentation" = "https://nrbnlulu.github.io/strawberry-django-auth/"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
force_sort_within_sections = true

[tool.mypy]
plugins = ["mypy_django_plugin.main", "strawberry.ext.mypy_plugin"]
ignore_missing_imports = "True"

[tool.django-stubs]
django_settings_module = "testproject.settings"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
